Rebase rec-ARM: * Update to latest files * Rename rec-ARM to arm_emitter
