+++ Makefile.common
@@ -403,15 +403,15 @@ endif
 # Recompiler (ARM)
 ifeq ($(WITH_DYNAREC), arm)
 	DYNAREC_USED = 1
-	SOURCES_CXX += $(CORE_DIR)/core/rec-ARM/rec_arm.cpp
-	SOURCES_ASM += $(CORE_DIR)/core/rec-ARM/ngen_arm.S
+	SOURCES_CXX += $(CORE_DIR)/core/arm_emitter/rec_arm.cpp
+	SOURCES_ASM += $(CORE_DIR)/core/arm_emitter/ngen_arm.S
 endif
 
 # Recompiler (ARM64)
 ifneq ($(filter $(WITH_DYNAREC), arm64 aarch64),)
 	DYNAREC_USED = 1
-	SOURCES_ASM += $(CORE_DIR)/core/rec-ARM64/ngen_arm64.S
-	SOURCES_CXX += $(CORE_DIR)/core/rec-ARM64/rec_arm64.cpp \
+	SOURCES_ASM += $(CORE_DIR)/core/arm_emitter/ngen_arm64.S
+	SOURCES_CXX += $(CORE_DIR)/core/arm_emitter/rec_arm64.cpp \
 					$(DEPS_DIR)/vixl/code-buffer-vixl.cc \
 					$(DEPS_DIR)/vixl/compiler-intrinsics-vixl.cc \
 					$(DEPS_DIR)/vixl/cpu-features.cc \
