+++ core/arm_emitter/E_Extend.h
