+++ core/arm_emitter/E_Misc.h
