+++ core/arm_emitter/E_Multiply.h
