+++ core/arm_emitter/E_Parallel.h
