+++ core/arm_emitter/E_Special.h
