+++ core/arm_emitter/H_psuedo.h
