+++ core/arm_emitter/H_state.h
