+++ core/arm_emitter/arm_coding.h
