+++ core/arm_emitter/arm_emitter.h
