+++ core/arm_emitter/arm_registers.h
