+++ core/arm_emitter/ngen_arm.S
